updated delete endpoint

+++ front-end/src/components/Delete.jsx
@@ -3,6 +3,7 @@ import React, { useState } from 'react';
 const Delete = () => {
   const [formData, setFormData] = useState({
     email: '',
+    secanswer: ''
   });
 
   const [errorFound, setErrorFound] = useState('');
@@ -18,6 +19,13 @@ const Delete = () => {
       return;
     }
 
+    const checkResponse2 = await fetch(`http://localhost:3000/checkAnswer?secanswer=${formData.secanswer}`);
+
+    if (checkResponse2.status !== 200) {
+      setErrorFound('Wrong Answer');
+      return;
+    }
+
     // If email exists, proceed with the deletion
     const deleteResponse = await fetch('http://localhost:3000/delete', {
       method: 'DELETE',
@@ -29,7 +37,7 @@ const Delete = () => {
 
     if (deleteResponse.status === 200) { //delete on proper server response (done to connect to backend)
       console.log('User deleted successfully');
-      setErrorFound('');
+      setErrorFound('User Deleted Successfully');
     } else {
       console.error('User deletion failed');
       setErrorFound('User deletion failed');
@@ -49,6 +57,13 @@ const Delete = () => {
           type="email"
           placeholder='Email'
           name="email"
+          required
+        />
+        <input
+          onChange={handleChange}
+          type="text"
+          placeholder='Security Answer'
+          name="secanswer"
           required
         />
         <input type="submit" value="Delete User" />
